feat: separate Product into component
